backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
aiosqlite==0.19.0
python-dotenv==1.0.0
twilio==8.10.0
requests==2.31.0
webauthn==1.11.1
cryptography==41.0.7
